Refactor db access for requests
